Update arguments of gadget and SPHDataset initializers
This removes old keyword arguments that are no longer supported (n_ref,
over_refine_factor, and index_ptype) and exposes two new keyword arguments to
control the creation of the bitmask index: index_order and index_filename.

index_order is a tuple that corresponds to the maximum level of the coarse and
refined morton index. For now it must be a two-element tuple or an integer. If
it's an integer then the order of the refined index is set to 1 (i.e. a trivial
refined index).

index_filename is a user-controllable base name for the particle index file.

This also removes all of the keyword arguments for _initialize_index, replacing
them with keyword arguments controlled by the user when creating a dataset.

--HG--
branch : yt