Hanlde invalid tipsy aux files
Right now if a file matching the tipsy aux file glob pattern is present in the
output directory, we attempt to parse the file as a valid tipsy aux file
die with a ValueError. Instead, let's detect this case and remove the field
from the list of detected tipsy auxiliary fields.

This was found in the context of swithing between mainline and the demeshening
branch. In the demeshening we create sidecar files (in particular, one named
`galaxy.00300.kdtree`) that are detected by mailine yt as auxiliary
files. However, this would happen for users as well if they happen to have a
file with a name matching the aux file glob pattern we use in their output
directory.