Fix the name of the colors keyword to matplotlib contour
The annotate_contour callback passes 'color' as a keyword to contour
from matplotlib, but this is not a known keyword. This appears to be
a typo in commit 9a2f4ff4a5; prior to then, the correct keyword (colors)
was passed to the contour plot. This commit will have the effect of
making all contours be black in color by default.