Use a level-by-level approach
Now loops on cells from the finest to the coarsest level. For each
level, we find all particles that within dx_level of one cell. For
these particles, we compute the actual distance and compare it the the
cell size (so that it should work for non-squared cells as
well). Particles within a cell are flagged and the code proceeds to
the next level.