Fix answer testing plugin
With this fix, answer name could be specified in the answer test case
itself. As a  result, all the answer tests could be executed by just one
nosetests command. Earlier, this was not possible and one nose command
had to be run per answer test as --answer-name had to be specified for
each answer test.

In addition, prepend py-version before AnswerTestingTest.answer_name,
to avoid conflict in answer-store images