Add feature to download test data
This feature helps in testing with real dataset, if the dataset is
not present in local repository. If the test dataset is already
present, this will not do anything. If test dataset is not
present, it will check the test data cache for corresponding tarfile.
If tarfile is present, it will extract it to the test data directory
location, otherwise it will download and extract the tar file.