Do not assume dicts are ordered
In Python < 3.6, dicts are not ordered so that iterating over key, values is not predictible. This caused the position to be filled with garbage (depending on the fields being read from disk) and then rejected by the selector, so that the shape of the data was depending on what was being read from disk (which influenced the order of the dictionary, which changed which garbage ends up in x, y, z).