Updated the requires_backend wrapper.
The requires_backend meta-wrapper in yt/testing.py was causing an error during
the execution of the unit tests under pytest. This is because, if the desired
backend was not found, then a lambda: None was returned, which caused an error
with the number of arguments received vs. the number of arguments expected for
the tests test_make_colormap and test_show_colormaps. Now, if the desired backend
is not found, a function called skip is returned by the meta-wrapper. Skip
prints a message about being unable to find the desired backend and then tells
pytest to skip the test.