Added pytest.ini
Right now, without the pytest-specific answer tests being implemented in yt,
using pytest causes the answer tests to be collected and run when trying to
only execute the unit tests. As such, this file has been created in order to
tell pytest to skip the frontend answer tests. This file should be removed
once all of the pytest-specific answer tests have been merged.