Added the _called_from_pytest variable.
Certain aspects of testing are handled differently by pytest and nosetests. As
such, the _called_from_pytest variable has been set to False in yt/__init__.py.
This variable is set to True in pytest_configure in the root conftest file.
This variable can be used to determine which test framework is being used and
then act accordingly. This is only useful while the switch to pytest is ongoing.