Implemented _called_from_pytest in two places.
If requires_backend could not find the passed backend then lambda : None was
returned. Howevever, this caused an error when using pytest. Defining a function
to return None does work, but it causes pytest to mark the test as passed, which
feels wrong, since the test wasn't actually run. However, calling pytest.skip
in order to flag the test as skipped causes the test to be broken out of immediately.
This causes an issue in test_requires_backend where one of the test cases is to
make sure that the proper None value was returned in the case of an unfound
backend. This issue prevents the assertion from ever occurring, since the test
is broken out of as soon as the pytest.skip is encountered or pytest.skip.Exception
is raised. Right now, that comparison is last and, if it succeeds, then the
test is marked as skipped. See the in-code comments. That should be fixed.