Fixed some flake8 issues.
Fixed a flake8 issue.

Fixed issues with accumulators tests.

There was an error loading the IsolatedGalaxy dataset for two of the
accumulators tests. Added those to the list of answer tests in tests.yaml
and decorated them with requires_ds. Also switched from calling yt.load
to data_dir_load, as is done with other answer tests.

Fixed a flake8 issue.