Fixed several bugs.
For amrvac, parametrizing the fields for testing requires either reading in the
data or else manually typing out every single field for each test dataset. The
issue is that pytest needs to generate the tests before it can decide whether
or not the test can be skipped. As such, if the amrvac data isn't present, even
if those tests are not being run, the test will fail since loading the data
will fail during test generation. I added a skip mark to those tests if the
data isn't found. This works, but is not satisfactory because if one is running
just those tests and all of the data isn't present (but some is), then, unless
one knows exactly how many tests there are supposed to be, nothing will fail,
resulting in a false positive testing run.

There was also a bug in my parametrization of the arepo tests. I had been using
"max" instead of "c" for the ds_obj parametrization.

There was also a bug in the arepo tests that called sph_answer. I was adding
"sph_answer" as an erroneous key to the hashes dictionary, which caused the
actual act of hashing to fail. It needed to be the same format as small_patch_amr,
which it now is.