Switched CI config files to use pytest.
What was previously pytest.ini has been split into two files, following the
nose example: pytest_answer.ini and pytest_unit.ini for the answer and unit
tests, respectively. Those files contain the options to pass to pytest,
including which files to ignore. For now, the still-existing nose frontend
answer tests are skipped as well as all of the non-frontend answer tests that
have not yet been converted to pytest.

pytest_modifyitems has been, well, modified. It now has a check for if the
test is not an answer test and the --with-answer-testing option is set, which
makes it so that unit tests are only run when --with-answer-testing is not set
and they are skipped when it is. Likewise, the answer tests are only run when
--with-answer-testing is set.