Ran port_files.
This renames all of the original nose test files to have a _nose suffix
and renames all of the _pytest files to the original file names (no suffix).
This was done to keep the original file history. It also adds nose back into
appveyor and travis so that it can be run side-by-side with pytest. This
results in the _answer and _unit files having to be modified so that nose
now ignores all of the normally named pytest files and pytest now ignores all
of the _nose files. The easiest way to do this on the nose side was to
introduce a new, temporary, config file called nose_ignore.cfg. This file
should be removed once the conversion to pytest is complete and nose is no
longer used. This switch also results in the pytest answer files losing the
_pytest suffix, so those answers need to be regenerated (hence the addition
of the --answer-store and --force-overwrite options to pytest_answer.ini).