Changes to testing utilities.
When saving the raw answers to an hdf5 file, the group being created
has the dataset name in it. For some datasets, this name includes the
full directory structure (e.g., IsolatedGalaxy/galaxy0030/galaxy0030).
The / causes erroneous subgroups to be created in the file so these
slashes were changed to underscores.

The second change made is to pad the results of the pr test so that the
raw array isn't ragged and can therefore be saved to an hdf5 file.