Store the Enzo-E version number in Enzo-E object
More recent versions of Enzo-E record the version number in the outputs. This change now stores that in the parameters attribute of EnzoEDataset instances (the value is associated with 'version_string').

This change also updates the check for whether Enzo-E particle masses should be interpretted as masses or densities to always interpret them as masses if a version string is specified. The previous way Enzo-E specified this choice (i.e. by appending a parameter called "mass_is_mass" to the output copy of the parameter file) ended up causing some problems on certain platforms.