Updated _is_valid
Newer (post Apr 2021) versions of gizmo export GIZMO_version=year_of_last_commit as a header variable. We should use this in preference to checking the gas metallicity, which might not exist. Since the names of the omegas have changed, we don't want to fall back to the gadget loader in the comoving case (in the non-cosmo case it shouldn't matter?)