minor bugfix in cholla frontend
Previously, the ``domain_right_edge`` attribute of ``ChollaDataset`` was
initialized by assigning it the values of the ``"domain"`` attribute
stored in the hdf5 file. This patch adjusts the logic so that the
``domain_right_edge`` attribute is now the sum of that array and the
array stored in the ``domain_left_edge`` attribute.