Better default behavior when cholla doesn't write mu attribute
Basic background: the yt frontend always tries the ``"mu"`` attribute
from Cholla hdf5 files to set the mean molecular weight. However, this
is not a standard attribute written by Cholla. When this attribute is
missing, the frontend falls back to a value of ``1.0`` (which is almost
always wrong).

It's on my todo list to start having Cholla write out the ``"mu"``
attribute (when not using species fields).

This patch tweaks the frontend to achieve better behavior for all
existing Cholla datasets without the ``"mu"`` attribute. In particular,
the tweaks enusre that we don't try to define a temperature field
unless ``"mu"`` attribute is explicitly defined (or the
``default_species_fields`` keyword argument is specified in ``yt.load``)