improves heuristic in default implementation of `GridIndex._chunk_io`
This commit improves the heuristic used by `GridIndex`’s default implementation of the `_chunk_io` method (with the `"auto"` chunk-sizing strategy) for determining how many grids to read in a given iteration.

For context, the heuristic was historically hardcoded to a value of 1000 grids. This works well for AMR simulations with small grids (e.g. 16^3, 32^3), but the heuristic is problematic when you have unified skmulations

This commit adopts a heuristic that tries to limit the number of grids in order to make sure we don't run out of memory.