Parallel reduce which domains should be read
For RAMSES datasets, each MPI rank may have already indexed fully
part of the domains, leading to different chunks across MPI ranks.

This means that we can do an MPI reduction over which domains
should be read to get the most specific subset of domains, but also
to ensure that all ranks have the same list of domains when chunking.